requirements.txt
streamlit
yfinance
pandas
numpy
matplotlib
plotly
tensorflow-cpu
ta==0.10.2 
scikit-learn